2022-12-8 16:51 完成trade从form表接受提交

+++ src/main/java/com/atguigu/mybatisplus/entity/Tradehistory.java
@@ -22,7 +22,7 @@ public class Tradehistory implements Serializable {
 
     private String name;
 
-    private Integer guessprice;
+    private float guessprice;
 
     private String content;
